Add web authentication session protocol
